Uploaded files
Initial setup, which includes general environment aliases to be sourced. A "job in" bash script, create job/shot bash scripts. Very simple, but useful to have. Also includes the /config/3DSceneColors.dark file that can be thrown in the $HOME/<HoudiniVersion>/config/ to override dark theme to not be just black at 0,0,0.